pyproject.toml
[tool.poetry]
name = "lazy-rabbit-agent"
version = "0.1.0"
description = "Lazy Rabbit Agent"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "backend"}, {include = "example"}]

[tool.poetry.dependencies]
python = "^3.12"
instructor = "*"
openai = "*"
fastapi = "*"
uvicorn = "*"
loguru = "0.7.2"
pymgclient = "1.3.1"
requests = "2.31.0"
python-dotenv = "1.0.1"
python-jose = "*"
python-multipart = "*"
passlib = "*"
websockets = "*"
cachetools = "5.5.0"
bcrypt = "4.0.1"
celery = "5.4.0"
redis = "5.0.8"
flower = "2.0.1"
langchain = "*"
langchain-community = "*"
langchain-openai = "*"
langchain_chroma = "*"
langchain_ollama = "*"
PyYAML = "6.0.2"
Jinja2 = "3.1.4"
prometheus-fastapi-instrumentator = "7.0.0"
neo4j = "5.24.0"
PyMySQL = "1.1.1"
alembic = "1.13.2"
wtforms = "3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"